Correction bad measures on 1st and last thread
Modification of measures, not by rdtsc but by couting the number of adds made

+++ include/PPN-microbench/cpu_frequency.hpp
@@ -12,7 +12,7 @@ class CPUFrequency : Microbench {
         int nbThreads;
         int nbMeasures = 0;
         int coresToExecute = 0;
-        std::vector<std::vector<u64>> measures;
+        std::vector<std::vector<double>> measures;
         std::vector<std::vector<u64>> benchTimes;
         void executeAdds();
